// libs/cdk/date/date-suggestion/domain/enums/default-date-interval-suggestion.ts
export const enum DefaultDateIntervalSuggestion {
    TODAY = 'TODAY',
    YESTERDAY = 'YESTERDAY',
    CALENDAR_WEEK = 'CALENDAR_WEEK',
    LAST_3_DAYS = 'LAST_3_DAYS',
    LAST_7_DAYS = 'LAST_7_DAYS',
    LAST_60_DAYS = 'LAST_60_DAYS',
    FIRST_DAY_OF_INTERVAL = 'FIRST_DAY_OF_INTERVAL',
    LAST_180_DAYS_OF_INTERVAL = 'LAST_180_DAYS_OF_INTERVAL',
}